lab/pyproject.toml
[project]
name = "signals-lab"
version = "0.1.0"
description = "Interactive research notebooks for trading signal exploration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marimo>=0.7.9",
    "polars>=1.0.0",
    "psycopg[binary]>=3.2.0",
    "plotly>=5.24.0",
    "altair>=5.3.0",
    "python-dotenv>=1.0.0",
    "trading-signals-pipeline",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
package = true

[tool.uv.sources]
"trading-signals-pipeline" = { path = "../pipe", editable = true }
